backend/main-maf-chat/requirements.txt
fastapi==0.133.0
uvicorn[standard]==0.41.0
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.13.1
agent-framework-core==1.0.0rc1
agent-framework-a2a==1.0.0b260219
mcp==1.24.0
a2a-sdk==0.3.5
aiosqlite==0.21.0
aiohttp==3.13.3
azure-identity==1.25.1
azure-monitor-opentelemetry==1.8.2
pypdf==5.4.0
python-multipart==0.0.20
